requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
coverage==5.5
crispy-bootstrap5==2024.2
cryptography==42.0.5
cycler==0.12.1
Deprecated==1.2.14
deprecation==2.1.0
Django==4.2.7
django-crispy-forms==2.1
dnspython==2.5.0
eyed3==0.9.7
filetype==1.2.0
fonttools==4.49.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
mutagen==1.47.0
mysql-connector==2.2.9
numpy==1.26.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
plotly==5.19.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pygame==2.5.2
pygame-widgets==1.1.5
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
toml==0.10.2
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.0
vonage==3.13.0
vonage-jwt==1.1.0
wrapt==1.16.0
